+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Fixed
+
+- Define undefined `XDGDIRS_VER`
+
 ## [1.0.4] - 2020-12-22
 
 ### Fixed