+++ CHANGELOG.md
@@ -7,9 +7,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [1.0.5] - 2022-02-18
+
 ### Fixed
 
 - Define undefined `XDGDIRS_VER`
+- Return value of `xdgDirs_init()`
 
 ## [1.0.4] - 2020-12-22
 
@@ -66,7 +69,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Documentation
 - README, CHANGELOG and AUTHORS
 
-[unreleased]: https://github.com/Jorengarenar/libXDGdirs/compare/v1.0.4...HEAD
+[unreleased]: https://github.com/Jorengarenar/libXDGdirs/compare/v1.0.5...HEAD
+[1.0.5]: https://github.com/Jorengarenar/libXDGdirs/compare/v1.0.4...v1.0.5
 [1.0.4]: https://github.com/Jorengarenar/libXDGdirs/compare/v1.0.3...v1.0.4
 [1.0.3]: https://github.com/Jorengarenar/libXDGdirs/compare/v1.0.2...v1.0.3
 [1.0.2]: https://github.com/Jorengarenar/libXDGdirs/compare/v1.0.1...v1.0.2